Создал и наполнил классы: FilmScrenings, Astract, ConsoleMessages
