+++ ConsoleApp2/Classes/Film.cs
@@ -34,5 +34,9 @@ namespace ConsoleApp2.Classes
             Console.WriteLine("Год выхода: {0}",Year);
             Console.WriteLine("Описание: {0}\n",Description);
         }
+        public static Film ChooseFilm(List<Film> films)//повтор методов вопрос что делать
+        {
+            return Abstract.ChooseEl(films);
+        }
     }    
 }
